Moved `LabelMap` implementation to `gen` package
